♻️[Refactor] 기존 Occasion, SpecialOption을 Filter enum에 통합
