+++ src/main/java/com/eventory/server/domain/product/entity/enums/FilterType.java
@@ -1,0 +1,30 @@
+package com.eventory.server.domain.product.entity.enums;
+
+public enum FilterType {
+
+    ;
+
+    public enum Occasion {
+        BIRTHDAY,
+        ANNIVERSARY,
+        PROPOSAL,
+        PARTY,
+        FIRST_BIRTHDAY,
+        GRADUATION,
+        PARENTS,
+        BRIDAL_SHOWER,
+        GENDER_REVEAL
+    }
+
+    public enum SpecialOption {
+        HANDMADE,
+        CUSTOM_MADE,
+        ENGRAVING,
+        PHOTO_PRODUCT,
+        MESSAGE_CARD,
+        GIFT_WRAP,
+        CUSTOMIZABLE,
+        GIFT_SET
+    }
+
+}
