--- src/main/java/com/eventory/server/domain/product/entity/enums/Occasion.java
@@ -1,13 +1,0 @@
-package com.eventory.server.domain.product.entity.enums;
-
-public enum Occasion {
-    BIRTHDAY,
-    ANNIVERSARY,
-    PROPOSAL,
-    PARTY,
-    FIRST_BIRTHDAY,
-    GRADUATION,
-    PARENTS,
-    BRIDAL_SHOWER,
-    GENDER_REVEAL
-}
