--- src/main/java/com/eventory/server/domain/product/entity/enums/SpecialOption.java
@@ -1,12 +1,0 @@
-package com.eventory.server.domain.product.entity.enums;
-
-public enum SpecialOption {
-    HANDMADE,
-    CUSTOM_MADE,
-    ENGRAVING,
-    PHOTO_PRODUCT,
-    MESSAGE_CARD,
-    GIFT_WRAP,
-    CUSTOMIZABLE,
-    GIFT_SET
-}
